sensors sgp30 and scd4x added for display, other changes

+++ drivers/mcp9808/mcp9808.h
@@ -1,4 +1,6 @@
 
-extern int  sensor_in;
+#include <stdint.h>
+
+extern uint16_t sensor_in;
 int mcp9808_open(void);
 float mcp9808_read(void);
